requirements.txt
fastapi
uvicorn[standard]
pydantic 
motor 
beanie 
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart
pydantic-settings
pydantic[email]
celery
redis
dotenv